Lee mi funciòn de validate pero tengo detalle con el status
